fix(options): Escape user and password if they have special characters
